community: generate sig-list.md
Creates an overview over all the sigs that are defined via sigs.yaml
Also corrects a couple of labels for the sigs.
