+++ Makefile
@@ -1,0 +1,5 @@
+generate:
+	go run ./generators/cmd/sigs
+
+validate-sigs:
+	go run ./validators/cmd/sigs
