+++ generators/cmd/sigs/README.md
@@ -1,0 +1,11 @@
+# sigs generator
+
+Generates markdown pages for the sigs, user groups and committees defined in `sigs.yaml`.
+
+Run it whenenver you have changed the [sigs.yaml](/sigs.yaml) 
+
+## Usage
+
+```bash
+make generate
+```
