+++ generators/cmd/sigs/sig-list.gomd
@@ -1,0 +1,65 @@
+{{- /*
+    This file is part of the KubeVirt project
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+        http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+
+    See the License for the specific language governing permissions and
+    limitations under the License.
+
+    Copyright the KubeVirt Authors.
+
+    */
+-}}
+{{- /* gotype: kubevirt.io/community/generators/cmd/sigs.SigListTemplateData */ -}}
+{{ define "group" }}|[{{ .Name }}]({{ .Dir }}/charter.md) |[{{ .Label }}](https://github.com/kubevirt/kubevirt/labels/{{ .Label }})|{{ if .Leadership }}{{ range $index2, $chair := .Leadership.Chairs }}* [{{ $chair.Name }}](https://github.com/{{ $chair.Github }}), {{ $chair.Company }}<br/>{{ end }}{{ end }} |{{ if .Contact }}[Slack]({{ .Contact.Slack}})<br/> [Mailing List]({{ .Contact.MailingList }}){{ end }} |{{ range $index3, $meeting := .Meetings }}* {{ $meeting.Description }}: [ {{ $meeting.Day }} at {{ $meeting.Time }} {{ $meeting.TZ }} ({{ $meeting.Frequency }}) ]({{ $meeting.URL }})<br/> {{ end }} |{{ end }}
+<!---
+This is an autogenerated file!
+
+Please do not edit this file directly, but instead make changes to the
+sigs.yaml file in the project root.
+
+To understand how this file is generated, see github.com/kubevirt/community/tree/main/generator/cmd/sigs/README.md
+--->
+
+Most community activity is organized into Special Interest Groups (SIGs),
+Working Groups, and the community meeting.
+
+SIGs follow these [guidelines](GOVERNANCE.md) although each of these groups may operate a little differently
+depending on their needs and workflow.
+
+Each group's material is in its subdirectory in this project.
+
+{{ if $.Sigs.Usergroups }}
+### User Groups
+
+| Name | Label | Chairs | Contact | Meetings |
+|------|-------|--------|---------|----------|
+{{ range $ug := $.Sigs.Usergroups }}{{ template "group" $ug }}
+{{ end }}
+{{ end }}
+
+{{ if $.Sigs.Sigs }}
+### Special Interest Groups
+
+| Name | Label | Chairs | Contact | Meetings |
+|------|-------|--------|---------|----------|
+{{ range $sig := $.Sigs.Sigs }}{{ template "group" $sig }}
+{{ end }}
+{{end}}
+
+{{ if $.Sigs.Committees }}
+### Committees
+
+| Name | Label | Chairs | Contact | Meetings |
+|------|-------|--------|---------|----------|
+{{ range $committee := $.Sigs.Committees }}{{ template "group" $committee }}
+{{ end }}
+{{end}}
