+++ go.mod
@@ -1,5 +1,10 @@
 module kubevirt.io/community
 
-go 1.16
+go 1.22
 
-require gopkg.in/yaml.v3 v3.0.0
+require gopkg.in/yaml.v3 v3.0.1
+
+require (
+	github.com/sirupsen/logrus v1.9.3 // indirect
+	golang.org/x/sys v0.0.0-20220715151400-c0bba94af5f8 // indirect
+)
